fix: landingpage minor
