+++ resources/js/Pages/LandingPagePengumuman.svelte
@@ -1,0 +1,3 @@
+<script lang="ts">
+
+</script>
